requirements.txt
grpcio
grpcio-tools
websockets
numpy
matplotlib
flask
flask-socketio
protobuf